+++ tests/leb-freeze/test_fp.py
@@ -49,3 +49,33 @@ def test_aberrated_pupil_should_not_fail_if_missing_noll_indexes():
     zernike_coeffs = [0.3, 0.5, 0.3, 0.6, 0.8, 0.3, 0.1]
 
     Pupil.from_system_params(num_px=NUM_PX[0], zernike_coeffs=zernike_coeffs)
+
+
+def test_pupil_set_p(fake_pupil):
+    new_pupil_data = np.random.random(NUM_PX).astype(np.complex128)
+
+    fake_pupil.set_p(new_pupil_data)
+
+    # Zero out values outside the pupil radius
+    y, x = np.ogrid[
+        -new_pupil_data.shape[0] // 2 : new_pupil_data.shape[0] // 2,
+        -new_pupil_data.shape[1] // 2 : new_pupil_data.shape[1] // 2,
+    ]
+    mask = x**2 + y**2 > fake_pupil.pupil_radius_px**2
+    new_pupil_data[mask] = 0
+
+    assert np.allclose(fake_pupil.p, new_pupil_data)
+
+
+def test_pupil_set_p_wrong_size(fake_pupil):
+    new_pupil_data = np.random.random((NUM_PX[0] + 1, NUM_PX[1] + 1)).astype(np.complex128)
+
+    with pytest.raises(ValueError):
+        fake_pupil.set_p(new_pupil_data)
+
+
+def test_pupil_set_p_wrong_dtype(fake_pupil):
+    new_pupil_data = np.random.random(NUM_PX).astype(np.float32)
+
+    with pytest.raises(ValueError):
+        fake_pupil.set_p(new_pupil_data)